Basic working version.

+++ WaterSim2/cubeverts.h
@@ -40,7 +40,7 @@ vertex cubeVertices[] = {
 	{ {  1.0,  1.0, -1.0 }, 1.0, {  0.0,  0.0, -1.0 }, 0.0, 1.0 }
 };
 
-GLushort cubeElements[] = {
+GLuint cubeElements[] = {
 	0,  1,  3,
 	1,  2,  3,
 	4,  5,  7,
